# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "An API for getting the most relevant missing people info from LizaAlerts.",
        "title": "Locator API",
        "contact": {
            "name": "Sam Komarov",
            "url": "github.com/k0marov",
            "email": "sam@skomarov.com"
        },
        "version": "1.0"
    },
    "host": "sber.skomarov.com",
    "basePath": "/api/v1",
    "paths": {
        "/locator/missing": {
            "get": {
                "description": "Given a geo location, returns a list of people from LizaAlertAPI,\nordered by their relevance (calculated from distance and days missing) to provided location.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locator"
                ],
                "summary": "Get info about relevant missing people",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Geo location latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Geo location longitude",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/delivery.MissingPersonResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery.MissingPersonResponse": {
            "type": "object",
            "properties": {
                "photo_url": {
                    "type": "string"
                }
            }
        }
    }
}